Cleaned up sources

+++ Lab2/iSort.h
@@ -3,7 +3,7 @@
 
 
 typedef int (*comparePtr)(const void *key1, const void *key2);
-void insertionSort(void const *A, int n, int vSize, comparePtr compare);
+void insertionSort(void const *A, size_t n, size_t vSize, comparePtr compare);
 #endif
